moved info frame to subfolder
