+++ ecris/csd/viewer/gui/__init__.py
@@ -2,4 +2,4 @@ from .controls import PlotControls, FileListControls, FileList, ElementButtons
 from .plot import Plot
 from .menu import AppMenu
 from .diagnostic_window import DiagnosticWindow
-from .file_info_pane import FileInfoPane
+from .info_frame.file_info_pane import FileInfoPane
