+++ ecris/csd/viewer/gui/controls/controls.py
@@ -7,7 +7,7 @@ import ttkbootstrap as ttk
 from .elements import ElementButtons
 from .file_list import FileList
 from ..plot import Plot
-from ..file_info_pane import FileInfoPane
+from ..info_frame.file_info_pane import FileInfoPane
 
 class FileListControls(tk.Frame):
     def __init__(self, owner, file_list: FileList, *args, **kwargs):
