+++ ecris/csd/viewer/gui/controls/file_list.py
@@ -4,7 +4,7 @@ from typing import List
 import ttkbootstrap as ttk
 
 from ...files.csd_file import CSDFile, get_files
-from ..file_info_pane import FileInfoPane
+from ..info_frame.file_info_pane import FileInfoPane
 
 BLUE = "#5200FF"
 WHITE = "#FFFFFF"
