+++ ecris/csd/viewer/gui/info_frame/__init__.py
@@ -1,0 +1,1 @@
+from .file_info_pane import FileInfoPane
