+++ ecris/csd/viewer/gui/info_frame/file_info_pane.py
